marche mais erreur mais ajouter constructur dans le message 404;

+++ EFA-Back-end/src/main/java/com/example/efabackend/service/ChatServiceImpl.java
@@ -28,7 +28,7 @@ public class ChatServiceImpl implements ChatService {
 
     @Override
     public Chat addChat(Chat chat) throws ChatAlreadyExistException {
-        if (chatRepository.existsById(chat.getChatId())) {
+        if (chatRepository.existsById((long) chat.getChatId())) {
             throw new ChatAlreadyExistException();
         }
         chat.setChatId(sequenceGeneratorService.generateSequence(Chat.SEQUENCE_NAME));
